tools/python/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
click==8.1.7
python-dotenv==1.0.0
locust==2.17.0
faker==20.1.0





